backend/requirements.txt
flask
flask-cors
joblib
pandas
scikit-learn
openai==0.28
sounddevice 
scipy 
PyPDF2
python-dotenv
pycryptodome